Refactor `spark_retrieve_data_outside.py` to `google_health_SQL.py` for modularity and streamline MySQL data retrieval logic
- Remove unused methods and redundant functionality from `spark_retrieve_data_outside.py`.
- Move core MySQL data retrieval functionality to newly created `google_health_SQL.py`.
- Simplify class design by injecting `logger`, `DB`, and `spark` as parameters.
